work in progress
Former-commit-id: 574f1d3a7e84184f64fcb7c4d6651e7e84c118e8
Former-commit-id: 278f7d1968638d74d7ea3f0d64db2ed7e6d7d48c

+++ src/level.h
@@ -82,7 +82,7 @@ void level_set_is_frozen(levelp level, uint8_t val);
 uint8_t level_is_frozen(levelp level);
 void level_set_is_paused(levelp level, uint8_t val);
 void level_place_plant_pod(levelp);
-void level_place_explosion(levelp level, int32_t x, int32_t y);
+void level_place_explosion(levelp level, double x, double y);
 extern uint8_t game_over;
 
 extern uint8_t god_mode;
